# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a BigAgent API server.",
        "title": "BigAgent API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/bigagent/showdata": {
            "get": {
                "description": "根据UUID查询Agent并转发请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent管理"
                ],
                "summary": "搜索Agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent UUID",
                        "name": "uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v1/add": {
            "post": {
                "description": "新增Agent的配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent配置"
                ],
                "summary": "添加Agent配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证密钥",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Agent配置信息",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AgentConfigDB"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/v1/del": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent配置"
                ],
                "summary": "删除Agent配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证密钥",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v1/edit": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent配置"
                ],
                "summary": "修改Agent配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证密钥",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v1/get": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent配置"
                ],
                "summary": "查询Agent配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证密钥",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v1/login": {
            "post": {
                "description": "处理用户登录请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登录接口",
                "parameters": [
                    {
                        "description": "登录表单",
                        "name": "loginForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/v1/push": {
            "post": {
                "description": "向所有在线Agent推送指定配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent配置"
                ],
                "summary": "推送Agent配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证密钥",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "配置ID",
                        "name": "config_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/v1/push_host": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent配置"
                ],
                "summary": "下发指定主机的Agent配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证密钥",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "model.AgentConfigDB": {
            "type": "object",
            "properties": {
                "auth_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data_name": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "host": {
                    "description": "主机地址",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "path": {
                    "description": "请求路径",
                    "type": "string"
                },
                "port": {
                    "description": "端口号",
                    "type": "integer"
                },
                "protocol": {
                    "description": "网络协议，如http或https",
                    "type": "string"
                },
                "ranges": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "slot_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "times": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "token": {
                    "description": "下面为前端传入的各种认证数据",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "permissions": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "roleId": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}